Move Py4jCallbackConnectionCleaner to Streaming
